Move `HealthCheckMiddleware` tests among `WebSocketClient` tests
They test just internal type for WSClient and we want to have the file structure clear -> XXX + XXX_Tests
